refactor a shell

+++ mount.sh
@@ -1,6 +1,23 @@
 #!/bin/bash
+#################################################
+#
+# Usage:
+#   Making a filesystem from a block device 
+#   and mount it on provided mount point
+#
+# Arguments:
+#   $1: Mountpoint directory path
+#
+# Returns:
+#   0: Sucessfully mounted
+#   1: Provided Mountpoint already has been used 
+#  -1: There's no block device which can be used
+#
+##################################################
+
+
 IFS=$'\n'
-MOUNT_POINT=("/data01" "/data02" "/data03")
+MOUNT_POINT=$1
 FILE_SYSTEM="xfs"
 
 
@@ -18,22 +35,35 @@ partition_path=($(echo "${partition_path[*]}" | sort | uniq ) )
 both=( `{ echo "${device_path[*]}"; echo "${partition_path[*]}"; } | sort | uniq -d`)
 target_device=(`{ echo "${device_path[*]}"; echo "${both[*]}"; } | sort | uniq -u`)
 
-i=0
+mountpoint ${MOUNT_POINT}
+if [ $? == 0 ]; then
+  echo "## A filesystem already has been on provided mount point.##" >&2
+  exit 1
+fi
+
 for item in ${target_device[*]}; do
     parted -s -a optimal ${item} mklabel gpt
-    parted -s -a optimal ${item} -- mkpart ${FILE_SYSTEM}  0% 100%
-    device=${item}"1"
-
-    mkfs -t ${FILE_SYSTEM} ${device}
-    if [ ! -d ${MOUNT_POINT[${i}]} ]; then
-       sudo mkdir ${MOUNT_POINT[${i}]}
-    fi
-    mountpoint ${MOUNT_POINT[${i}]}
+    
     if [ $? != 0 ]; then
-       echo "mount on mountpoint"
-       uuid=`sudo blkid -o export ${device}  | grep ^UUID`
-       mount ${device} ${MOUNT_POINT[${i}]}
-       echo ${uuid} ${MOUNT_POINT[${i}]} ${FILE_SYSTEM} defaults,nofail 0 0 | tee -a /etc/fstab
+      echo "## A filesystem already has been used  ##"
+      echo "## Search next block device ##"
+      continue
+    else
+      parted -s -a optimal ${item} -- mkpart ${FILE_SYSTEM}  0% 100%
+      device=${item}"1"
+
+      mkfs -t ${FILE_SYSTEM} ${device}
+      if [ ! -d ${MOUNT_POINT} ]; then
+         mkdir ${MOUNT_POINT}
+      fi
+      echo "## mount a filesystem on a provided mountpoint ##"
+      uuid=`blkid -o export ${device}  | grep ^UUID`
+      mount ${device} ${MOUNT_POINT}
+      echo ${uuid} ${MOUNT_POINT} ${FILE_SYSTEM} defaults,nofail 0 0 | tee -a /etc/fstab
+      echo "## mount process has been successfully completed ##"
+      exit 0
     fi
-    i+=1
 done
+
+echo "There is no filesystem which can be mounted" >&2
+exit -1
